Move wsl recommendation into terminalContrib
Part of #230129
